+++ src/pychan/__init__.py
@@ -1,7 +1,7 @@
 from importlib_metadata import PackageNotFoundError, version
 
 try:
-    __version__ = version(dist_name)
+    __version__ = version(__name__)
 except PackageNotFoundError:
     __version__ = "unknown"
 finally: